+++ src/endpoints.rs
@@ -20,6 +20,7 @@ pub(crate) enum Endpoint {
     franchises,
     age_ratings,
     player_perspectives,
+    platform_logos,
 }
 
 pub(crate) fn get_endpoint_url(endpoint: &Endpoint) -> String {
@@ -41,5 +42,6 @@ pub(crate) fn get_endpoint_url(endpoint: &Endpoint) -> String {
         Endpoint::franchises => format!("{}/franchises", BASE_URL),
         Endpoint::age_ratings => format!("{}/age_ratings", BASE_URL),
         Endpoint::player_perspectives => format!("{}/player_perspectives", BASE_URL),
+        Endpoint::platform_logos => format!("{}/platform_logos", BASE_URL),
     }
 }